test:(repo-memory) separate shared job contracts from native launchers
